requirements.txt
flask
pyflakes
discord
pynacl
python-dotenv
waitress
